+++ src/components/Icon/icons/IconBell/index.js
@@ -1,0 +1,19 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'emotion';
+import { Icon, type GenericIconProps } from '../../Icon';
+import { colors } from '../../../../variables';
+import image from './icon-bell.svg';
+
+const styles = css`
+  width: 16px;
+  height: 16px;
+  fill: ${colors.dark65};
+`;
+
+export const IconBell = ({ className }: GenericIconProps) => (
+  <Icon
+    className={cx(styles, className)}
+    glyph={image}
+  />
+);