+++ CHANGELOG.md
@@ -20,8 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
   - IconEyeClosed
   - IconEyeOpened
+  - IconBell
 
-- Added AppHeader component.
+- Added AppLayout, AppHeader components.
 
 ## [0.23.0] - 2020-09-01